requirements.txt
scrapy
PyPDF2
pdf2image
pytesseract
openai
python-dotenv
requests
Pillow